Update __main__.py
Modified the process_lines method to set the global_time of the voice to zero, each time a new text is being generated. The variable global_time is used to keep track of the ms of the start and end time for the phoneme mapping, otherwise a new start time will be generated for each sentence. Then the filename that is being generated is lastly printed to the debug.